GSF.TimeSeries: Changed version numbers for configuration augmentation to unsigned long.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/gsf/Main;C89848

+++ Source/Libraries/Adapters/HistorianAdapters/LocalOutputAdapter.cs
@@ -793,7 +793,7 @@ namespace HistorianAdapters
         // Apply historian configuration optimizations at start-up
         // ReSharper disable once UnusedMember.Local
         // ReSharper disable once UnusedParameter.Local
-        private static void OptimizeLocalHistorianSettings(AdoDataConnection database, string nodeIDQueryString, int trackingVersion, string arguments, Action<string> statusMessage, Action<Exception> processException)
+        private static void OptimizeLocalHistorianSettings(AdoDataConnection database, string nodeIDQueryString, ulong trackingVersion, string arguments, Action<string> statusMessage, Action<Exception> processException)
         {
             // Make sure setting exists to allow user to by-pass local historian optimizations at startup
             ConfigurationFile configFile = ConfigurationFile.Current;
